requirements.txt
oemof.solph==0.4.4
